requirements.txt
# general modules
django >3.2,<3.3
python-decouple 
pillow
djangorestframework
markdown       
django-filter  
djangorestframework-simplejwt
djoser
drf-yasg[validation]


# email
django-mail-templated

# background process and cash
celery
redis
django-celery-beat
django-redis